GeneralMenu simplification.
This commit contains partial changes required to keep up with the
GeneralMenu simplification. It removes elements no longer necessary and
updates the comment to reflect the fact that there are now six submenu
entries instead of five.

To complete the change, the existing graphics for gmenu/gmenu_*.png
headings need to be modified so that they reflect the new columns. The
required filenames for the colunms are (left to right): 'gmenu_plugin',
'gmenu_media', gmenu_tv', 'gmenu_internet' and 'gmenu_task'

gmenu_up.png and gmenu_down.png do not require changes.

The remaining gmenu_*.png files can be removed.