Improve BackgroundImage and Spinner initialisation code
Change the code defauly from None to "" to better match the saved value.

Rescan the BackgroundImage and Spinner source directories every time the plugin is started to allow the available list of selections to be updated without requiring a full reboot.  Exiting and restarting the plugin should be enough to refresh the lists and pick up any recent changes.