[plugin.py] IMPORTANT: Correct significant spacing and code errors
This commit is to correct a significant merge malfunction that caused code indentation to be incorrectly changed.  The code in the two to three previous commits should NEVER be used or trusted.  The commits will be left in place to show the transition in features and functionality.  The code in this commit comes from a working configuration.