Rework NumericalTextInput overlay screen
This change makes the NumericalTextInput display more usable with all language and locale options.

The previous layout replicated the organisation of the number buttons on the remote control.  While this is an obvious layout it restricts the number of characters that could be assigned to each number.  The most apparent issue was that all the characters could not be displayed and hence the current character highlighting could not work.  The new layout is a simple list that allows all the characters to remain visible at all times.

The new design is also a little more screen efficient.