[skin_infobars.xml] Modify TimeshiftState screen to avoid crash
There is an instability in the "TimeshiftState" screen code such that if the "PTSeekBack" widget is defined and set to 0 width the code can crash.  The crash is eliminated if the widget is totally omitted.  The issue with this is that there is now an error when the screen is loaded that the "PTSeekBack" and "PTSeekPointer" widgets are undefined.

If the widgets are defined but shown outside of the screen window the code functionality changes!!!  If the widget is undefined the "0" button exits timeshift but the timeshift buffer is still active.  If the widget is defined the "0" button does nothing.

The undefined state offers users a more functional UI so the screen load errors will be ignored for now.