Update text formatting attributes
This change renames the "borderColor" and "borderWidth" attributes to "textBorderColor" and "textBorderWidth".  While the original attribute names remain functional, for the foreseen future, this change stands as an example of the more concise and accurate attribute names that better reflect the attribute functionality.  New skin functionality now allows the widget itself to also have border colors and widths "widgetBorderColor" and "widgetBorderWidth" respectively.  Other bolder color and widths are also soon to be implemented.  This rename will resolve potential ambiguity that may occur.